requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
jinja2==3.1.6
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.4.0
pytest==8.3.5
httpx==0.27.2